requirements.txt
pillow==11.0.0
python-dotenv
plotly==5.24.1
kaleido==0.2.1
igraph==0.11.8
python-magic==0.4.27
pymongo==4.10.1
python-magic-bin==0.4.14; platform_system == "Windows"
